Refactoring the projecto
Making some changes in names or imports
