+++ src/models/account-type.model.ts
@@ -1,0 +1,3 @@
+import { Types } from '../persistence/repositories/base';
+export interface AccountTypeModel extends Types{
+}
