--- src/models/accountType.model.ts
@@ -1,3 +1,0 @@
-import { Types } from '../persistence/repositories/base/';
-export interface AccountTypeModel extends Types{
-}
