+++ src/models/document-type.model.ts
@@ -1,0 +1,3 @@
+import { Types } from '../persistence/repositories/base';
+export interface DocumentTypeModel extends Types{
+}
