--- src/models/documentType.model.ts
@@ -1,3 +1,0 @@
-import { Types } from '../persistence/repositories/base/';
-export interface DocumentTypeModel extends Types{
-}
