+++ src/models/index.ts
@@ -1,6 +1,6 @@
+export * from './account-type.model';
 export * from './account.model';
-export * from './accountType.model';
 export * from './customer.model';
 export * from './deposit.model';
-export * from './documentType.model';
+export * from './document-type.model';
 export * from './transfer.model';
