+++ src/persistence/entities/document-type.entity.ts
@@ -1,4 +1,4 @@
-import { DocumentTypeModel } from '../../models/';
+import { DocumentTypeModel } from '../../models';
 import { v4 as uuid} from 'uuid';
 
 export class DocumentTypeEntity implements DocumentTypeModel{
