+++ src/persistence/entities/index.ts
@@ -1,7 +1,6 @@
+export * from './account-type.entity';
 export * from './account.entity';
-export * from './account_type.entity';
 export * from './customer.entity';
 export * from './deposit.entity';
-export * from './document_type.entity';
-export * from './transfer.entity';
+export * from './document-type.entity';
