+++ src/persistence/repositories/index.ts
@@ -1,8 +1,8 @@
+export * from './account-type.repository';
 export * from './account.repository';
-export * from './account_type.repository';
 export * from './base';
 export * from './customer.repository';
 export * from './deposit.repository';
-export * from './document_type.repository';
+export * from './document-type.repository';
 export * from './interfaces';
 export * from './transfer.repository';
