+++ src/persistence/repositories/interfaces/account-type-repository.interface.ts
@@ -1,4 +1,4 @@
-import { AccountTypeEntity } from '../../entities/account_type.entity';
+import { AccountTypeEntity } from '../../entities/account-type.entity';
 import { ICRUD } from './base/CRUD.interface';
 
 export interface AccountTypeRepositoryInterface extends ICRUD<AccountTypeEntity>{}
