+++ src/persistence/repositories/interfaces/document-type.repository.interface.ts
@@ -1,4 +1,4 @@
-import { DocumentTypeEntity } from '../../entities/document_type.entity';
+import { DocumentTypeEntity } from '../../entities/document-type.entity';
 import { ICRUD } from './base/CRUD.interface';
 
 export interface DocumentTypeRepositoryInterface extends ICRUD<DocumentTypeEntity>{}
